Keep sonar-home volume data as UID 1000 instead of root
docker compose will run sonarqube as root which will create the missing
data. In the end, sonarqube cannot access the directory structure
because sq is unable to access it as root after that.

This commit will keep .gitkeep files so the directory structure and the
associated user permissions are retained with UID:GID 1000